- Updated transform to work in a 2d environment better. - Updated so that the render function knows the delta time since it was last called. - Added basis for sprite class that should support still and animated textures.

+++ engine.cpp
@@ -112,7 +112,7 @@ namespace base {
 	{
 
 	}
-	void Engine::update(unsigned long dt)
+	void Engine::update(float dt)
 	{
 
 	}
